topfish rescale_layout_polar: push down 'size_t' changes
Allows removing a cast compromise from the earlier commit. This involves
introducing a lot of cast compromises but these will be removed in future
commits. Squashes 4 -Wsign-conversion warnings.

+++ cmd/smyrna/hier.c
@@ -91,7 +91,7 @@ void positionAllItems(Hierarchy * hp, focus_t * fs, reposition_t * parms)
 	switch (parms->rescale) {
 	case Polar:
 	    rescale_layout_polar(x_coords, y_coords, fs->x_foci,
-				 fs->y_foci, fs->num_foci, (int)counter,
+				 fs->y_foci, fs->num_foci, counter,
 				 interval, width, height, margin,
 				 distortion);
 	    break;
